requirements.txt
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
Pillow>=8.3.0
diffusers>=0.21.0
transformers>=4.21.0
accelerate>=0.20.0
scikit-learn>=1.1.0
tqdm>=4.64.0
pathlib2>=2.3.0